+++ src/main/java/school/faang/user_service/config/RedisConfig.java
@@ -4,10 +4,15 @@ import org.springframework.boot.CommandLineRunner;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.HashOperations;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.ZSetOperations;
+import org.springframework.data.redis.listener.ChannelTopic;
+import org.springframework.data.redis.listener.RedisMessageListenerContainer;
+import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
+import school.faang.user_service.redis.UserBanSubscriber;
 import school.faang.user_service.service.leaderboard.RedisWarmUpService;
 
 @Configuration
@@ -32,6 +37,30 @@ public class RedisConfig {
     @Bean
     public HashOperations<String, String, String> hashOperations(RedisTemplate<String, String> redisTemplate) {
         return redisTemplate.opsForHash();
+    }
+
+    @Bean
+    public RedisConnectionFactory redisConnectionFactory() {
+        return new LettuceConnectionFactory();
+    }
+
+    @Bean
+    public ChannelTopic userBanTopic() {
+        return new ChannelTopic("user_ban");
+    }
+
+    @Bean
+    public RedisMessageListenerContainer redisContainer(RedisConnectionFactory connectionFactory,
+                                                        MessageListenerAdapter listenerAdapter) {
+        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
+        container.setConnectionFactory(connectionFactory);
+        container.addMessageListener(listenerAdapter, userBanTopic());
+        return container;
+    }
+
+    @Bean
+    public MessageListenerAdapter listenerAdapter(UserBanSubscriber subscriber) {
+        return new MessageListenerAdapter(subscriber, "onMessage");
     }
 
     @Bean
