--- src/main/java/school/faang/user_service/redis/RedisConfig.java
@@ -1,37 +1,0 @@
-package school.faang.user_service.redis;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.listener.ChannelTopic;
-import org.springframework.data.redis.listener.RedisMessageListenerContainer;
-import org.springframework.data.redis.listener.adapter.MessageListenerAdapter;
-
-@Configuration
-public class RedisConfig {
-
-    @Bean
-    public RedisConnectionFactory redisConnectionFactory() {
-        return new LettuceConnectionFactory();
-    }
-
-    @Bean
-    public ChannelTopic userBanTopic() {
-        return new ChannelTopic("user_ban");
-    }
-
-    @Bean
-    public RedisMessageListenerContainer redisContainer(RedisConnectionFactory connectionFactory,
-                                                        MessageListenerAdapter listenerAdapter) {
-        RedisMessageListenerContainer container = new RedisMessageListenerContainer();
-        container.setConnectionFactory(connectionFactory);
-        container.addMessageListener(listenerAdapter, userBanTopic());
-        return container;
-    }
-
-    @Bean
-    public MessageListenerAdapter listenerAdapter(UserBanSubscriber subscriber) {
-        return new MessageListenerAdapter(subscriber, "onMessage");
-    }
-}
