+++ src/main/resources/application.yaml
@@ -1,7 +1,3 @@
-app:
-  leaderboard:
-    max-cached-size: 1000
-
 payment:
   api:
     url: http://localhost:9081/api/payment
@@ -12,6 +8,7 @@ spring:
     url: jdbc:postgresql://localhost:5432/postgres
     username: user
     password: password
+
   jpa:
     database-platform: org.hibernate.dialect.PostgreSQLDialect
     hibernate:
@@ -51,10 +48,15 @@ app:
     localhostUrlPrefix: "http://localhost:9000"
     fileExtension: ".png"
     contentType: "image/png"
+
   minio:
     avatars:
       endpoint: http://localhost:9000
       access-key: HZykibzYtPG1ABN2J073
       secret-key: qjGfUbiitoZyWPNYS24G4EQKDsxCba9rK28DoqEj
+
     bucket:
       name: "corpbucket"
+
+  leaderboard:
+    max-cached-size: 1000
